requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
openai==1.61.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
spitch==1.19.0
starlette==0.45.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
